+++ .idea/compiler.xml
@@ -10,4 +10,9 @@
       </profile>
     </annotationProcessing>
   </component>
+  <component name="JavacSettings">
+    <option name="ADDITIONAL_OPTIONS_OVERRIDE">
+      <module name="cinema-tickets" options="-parameters" />
+    </option>
+  </component>
 </project>
