--- cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/TicketService.java
@@ -1,10 +1,0 @@
-package uk.gov.dwp.uc.pairtest;
-
-import uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
-import uk.gov.dwp.uc.pairtest.exception.InvalidPurchaseException;
-
-public interface TicketService {
-
-    void purchaseTickets(Long accountId, TicketTypeRequest... ticketTypeRequests) throws InvalidPurchaseException;
-
-}
