--- cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/domain/TicketTypeRequest.java
@@ -1,29 +1,0 @@
-package uk.gov.dwp.uc.pairtest.domain;
-
-/**
- * Immutable Object
- */
-
-public class TicketTypeRequest {
-
-    private int noOfTickets;
-    private Type type;
-
-    public TicketTypeRequest(Type type, int noOfTickets) {
-        this.type = type;
-        this.noOfTickets = noOfTickets;
-    }
-
-    public int getNoOfTickets() {
-        return noOfTickets;
-    }
-
-    public Type getTicketType() {
-        return type;
-    }
-
-    public enum Type {
-        ADULT, CHILD , INFANT
-    }
-
-}
