--- cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/exception/InvalidPurchaseException.java
@@ -1,7 +1,0 @@
-package uk.gov.dwp.uc.pairtest.exception;
-
-public class InvalidPurchaseException extends RuntimeException {
-    public InvalidPurchaseException(String message) {
-        super(message);
-    }
-}
