--- cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/pricing/TicketPricingStrategy.java
@@ -1,8 +1,0 @@
-package uk.gov.dwp.uc.pairtest.pricing;
-
-import uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
-
-public interface TicketPricingStrategy {
-    int calculateTotalAmount(TicketTypeRequest... ticketTypeRequests);
-    int calculateTotalSeats(TicketTypeRequest... ticketTypeRequests);
-}
