--- cinema-tickets-java/src/main/java/uk/gov/dwp/uc/pairtest/validation/TicketValidator.java
@@ -1,8 +1,0 @@
-package uk.gov.dwp.uc.pairtest.validation;
-
-import uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
-import uk.gov.dwp.uc.pairtest.exception.InvalidPurchaseException;
-
-public interface TicketValidator {
-    void validate(Long accountId, TicketTypeRequest... ticketTypeRequests) throws InvalidPurchaseException;
-}
