+++ src/main/java/com/example/cinema_tickets/CinemaTicketsApplication.java
@@ -1,0 +1,13 @@
+package com.example.cinema_tickets;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CinemaTicketsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CinemaTicketsApplication.class, args);
+	}
+
+}
