+++ src/main/java/com/example/cinema_tickets/uk/gov/dwp/uc/pairtest/TicketService.java
@@ -1,0 +1,10 @@
+package com.example.cinema_tickets.uk.gov.dwp.uc.pairtest;
+
+import com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
+import com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.exception.InvalidPurchaseException;
+
+public interface TicketService {
+
+    void purchaseTickets(Long accountId, TicketTypeRequest... ticketTypeRequests) throws InvalidPurchaseException;
+
+}
