+++ src/main/java/com/example/cinema_tickets/uk/gov/dwp/uc/pairtest/exception/GlobalExceptionHandler.java
@@ -1,0 +1,23 @@
+package com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import com.example.cinema_tickets.dto.ApiResponse;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(InvalidPurchaseException.class)
+    public ResponseEntity<ApiResponse> handleInvalidPurchaseException(InvalidPurchaseException ex) {
+        ApiResponse response = new ApiResponse("FAILURE", ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ApiResponse> handleGeneralException(Exception ex) {
+        ApiResponse response = new ApiResponse("ERROR", "An unexpected error occurred: " + ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+}
