+++ src/main/java/com/example/cinema_tickets/uk/gov/dwp/uc/pairtest/exception/InvalidPurchaseException.java
@@ -1,0 +1,7 @@
+package com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.exception;
+
+public class InvalidPurchaseException extends RuntimeException {
+    public InvalidPurchaseException(String message) {
+        super(message);
+    }
+}
