+++ src/main/java/com/example/cinema_tickets/uk/gov/dwp/uc/pairtest/pricing/TicketPricingStrategy.java
@@ -1,0 +1,7 @@
+package com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.pricing;
+
+import com.example.cinema_tickets.uk.gov.dwp.uc.pairtest.domain.TicketTypeRequest;
+public interface TicketPricingStrategy {
+    int calculateTotalAmount(TicketTypeRequest... ticketTypeRequests);
+    int calculateTotalSeats(TicketTypeRequest... ticketTypeRequests);
+}
