+++ src/test/java/com/example/cinema_tickets/CinemaTicketsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.cinema_tickets;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CinemaTicketsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
